Design: See long desc (traits make the driver)
Starting with drivers over hardware that is a tight-coupling of 3 or 2 pins
we are trying a pattern where a driver is merely a trait impl of some sort.
E.g. A pin triplet of `SetDutyCycle` implementors can implement bldc motor control
or a pin triplet of `OutputPin` could implement a trait that sets phase states.